feat: hnsw implement

+++ benchmarks/include/benchmark_types.h
@@ -36,7 +36,8 @@ const size_t W2V_SIZE = 300;
 typedef de::EuclidPoint<double, W2V_SIZE> Word2VecRec;
 
 const size_t ANNSize = 128;
-typedef de::EuclidPoint<uint64_t, ANNSize> ANNRec;
+// typedef de::EuclidPoint<uint64_t, ANNSize> ANNRec;
+typedef de::EuclidPoint<uint8_t, ANNSize> ANNRec;
 
 struct euclidean_distance {
     double operator()(const Word2VecRec &first, const Word2VecRec &second) const {
